+++ CHANGELOG.rst
@@ -10,8 +10,10 @@ Unreleased
 ----------
 
 **Added**
+* Output nested printpoints  
 
 **Changed**
+* Updates in the visualization of PrintPoints
 
 **Fixed**